Add files via upload
The aim of project is to monitor the performance of each employee.
The name of our Project name is “Employee Monitoring System”. This project is made using connectivity with Oracle.
Our application is based on keeping the record of each and every employee in a single framework which is useful for many MNC’s.
i. Login:
The application will consist of two login windows:-
1.Admin Login Window
2. Employee Login Window

Admin Login Window:
A username and a password will be provided to the Administrative department of the company.
Employee Login Window:
A separate username and a password will be provided to the Employee members which are a part of any department in the company.
ii. Roles

Role of Admin:
 This is the main module mentioned above. This module has all control of the project.
 Let’s suppose there’s any project which is going to take place in any college eg.an app for messaging.
 The admin will login into his/her account and will be connected to the group page which will consists of various text fields like group id, task id, number of members
 The admin will be connected to the task page which will consists of various textfields like task id , task name, deadline , start date, points and description.
 Each and every control of data entry, data deletion and data updation is taken care by Admin module.
Roles of Employee:
 This is the second module of project which has further sub menus.
 Employee will login through their login page which consists of the two two text fields username and password.
 Employee table consists of various text fields emp id , emp name, dept id, designation , salary , hiredate , sex, emp_country , contact info, marital status , group id, task id, email.